Made changes to PartnershipCode5015.java
